refactor(server): merge text_overlay into composition
